small cleanup

+++ crates/cubecl-linalg/src/matmul/cmma_instruction/base.rs
@@ -60,9 +60,7 @@ macro_rules! impl_matmul_instruction {
             (I, O): CmmaValid<I, O>,
         {
             fn can_process(problem: MatmulProblem) -> bool {
-                problem.m as u32 == Self::M
-                    && problem.n as u32 == Self::N
-                    && problem.k as u32 == Self::K
+                problem.m == Self::M && problem.n == Self::N && problem.k == Self::K
             }
 
             fn requirements(_problem: MatmulProblem) -> Requirements {
